+++ utils/game/validateBoard.ts
@@ -1,8 +1,8 @@
 /* 
   Validate board against /api/check endpoint
 */
-import CheckBoardRequestData from '../pages/api/types/CheckBoardRequestData'
-import CheckBoardResponseData from '../pages/api/types/CheckBoardResponseData'
+import CheckBoardRequestData from '../../pages/api/types/CheckBoardRequestData'
+import CheckBoardResponseData from '../../pages/api/types/CheckBoardResponseData'
 
 const validateBoard = async (validateBoardData: CheckBoardRequestData) => {
   const response = await fetch('http://localhost:3000/api/check', {
